Rename fusion files
