+++ onnxscript/rewriter/ort_fusions/_core.py
@@ -17,8 +17,8 @@ from onnxscript.rewriter.ort_fusions.attention import fuse_attention
 from onnxscript.rewriter.ort_fusions.bias_gelu import fuse_bias_gelu
 from onnxscript.rewriter.ort_fusions.cos_sin_cache import fuse_cos_sin_cache
 from onnxscript.rewriter.ort_fusions.erfgelu import fuse_erfgelu
-from onnxscript.rewriter.ort_fusions.fuse_mha_bias import fuse_mha_bias
-from onnxscript.rewriter.ort_fusions.fuse_packed_qkv_gqa import fuse_qkv_gqa
+from onnxscript.rewriter.ort_fusions.mha_bias import mha_bias
+from onnxscript.rewriter.ort_fusions.packed_qkv_gqa import fuse_qkv_gqa
 from onnxscript.rewriter.ort_fusions.gelu import fuse_gelu
 from onnxscript.rewriter.ort_fusions.gqa import fuse_gqa
 from onnxscript.rewriter.ort_fusions.mha import fuse_mha1, fuse_mha2
@@ -97,7 +97,7 @@ def fuse_xformers(model: ir.Model, debug: bool = False) -> tuple[ir.Model, dict[
         fusion_count["mha_bias"] = 0
         fusion_count["attention"] = 0
     else:
-        fusion_count["mha_bias"] = fuse(fuse_mha_bias)
+        fusion_count["mha_bias"] = fuse(mha_bias)
         fusion_count["attention"] = fuse(fuse_attention)
     fusion_count["gelu"] = fuse(fuse_gelu)
     fusion_count["bias_gelu"] = fuse(fuse_bias_gelu)
