+++ onnxscript/rewriter/ort_fusions/mha_bias.py
@@ -163,7 +163,7 @@ class FuseBiasMHA(pattern.RewriteRuleClassBase):
         )
 
 
-fuse_mha_bias_rules = pattern.RewriteRuleSet([FuseBiasMHA.rule()])
+mha_bias_rules = pattern.RewriteRuleSet([FuseBiasMHA.rule()])
 
 
-fuse_mha_bias = _fusion_utils.apply_fusion_rules(fuse_mha_bias_rules)
+mha_bias = _fusion_utils.apply_fusion_rules(mha_bias_rules)
